# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Learning Core API",
    "version": "1.0.4",
    "description": "Headless multi-tenant learning API: structured course content, recomputed progress, server-side scoring with partial credit, publicly verifiable credentials, and metered usage. Content is typed JSON blocks, never markup."
  },
  "servers": [
    {
      "url": "http://167.233.117.23:8461"
    }
  ],
  "paths": {
...
    "/v1/verify/{ref}": {
      "get": {
        "summary": "Public credential verification. No API key. Exact-match only; strictly rate limited per IP. Fields: public_ref, status (active|revoked), learner_name, course_title, rank_name, rank_meta (opaque tenant JSON, e.g. a display colour), issuer, issued_at, revoked_at, revoke_reason.",
        "operationId": "get_v1_verify_ref",
        "security": [],
        "parameters": [
          {
            "name": "ref",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The credential public record.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "default": {
            "description": "Problem Details",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "apiKey": {
        "type": "http",
        "scheme": "bearer",
        "description": "Tenant API key, e.g. reli_xJ9w… Scopes: read, write, admin."
      }
    },
    "schemas": {
      "Problem": {
        "type": "object",
        "description": "RFC 9457 Problem Details. Every type URL resolves to a hosted page.",
        "required": [
          "type",
          "title",
          "status"
        ],
        "properties": {
          "type": {
            "type": "string",
            "format": "uri"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string"
          },
          "request_id": {
            "type": "string"
          }
        }
      }
    }
  }
}